contextual-rag/requirements.txt
exllamav2
transformers
sentence-transformers
torch
elasticsearch
numpy
pandas
matplotlib
nltk
tqdm
huggingface-hub
langchain
langchain-experimental
langchain-community
langchain-core
tokenizers
requests
bert-score
rouge-score
semantic-router
semantic-chunkers
pydantic
python-dotenv
datasets